+++ api/services/services.account.js
@@ -1,6 +1,6 @@
 const { account, user, branch, transaction } = require('../models/models.index')
 const accountMapper = require('../mappers/mappers.account')
-const ErrorGeneric = require('../utils/errors/erros.generic_error')
+const ErrorGeneric = require('../utils/errors/erros.generic-error')
 
 const listAllAccountService = async () => {
   try {
