+++ api/services/services.bank.js
@@ -1,6 +1,6 @@
 const { bank } = require('../models/models.index')
 const bankMapper = require('../mappers/mappers.bank')
-const ErrorGeneric = require('../utils/errors/erros.generic_error')
+const ErrorGeneric = require('../utils/errors/erros.generic-error')
 
 const listAllBanksService = async () => {
   try {
